fix(error)!: Group error items
I think I was worried about these being too tied to `Parser` in #13 but really,
this is more tied to `ParseError` and the other error docs.

BREAKING CHANGE:
- Moved `ErrMode` and `Needed` out of the root namespace
