+++ examples/s_expression.rs
@@ -352,7 +352,7 @@ fn eval_expression(e: Expr) -> Option<Expr> {
 /// us call eval on a string directly
 fn eval_from_str(src: &str) -> Result<Expr, String> {
   parse_expr(src)
-    .map_err(|e: winnow::ErrMode<VerboseError<&str>>| format!("{:#?}", e))
+    .map_err(|e: winnow::error::ErrMode<VerboseError<&str>>| format!("{:#?}", e))
     .and_then(|(_, exp)| eval_expression(exp).ok_or_else(|| "Eval failed".to_string()))
 }
