+++ src/bits/complete.rs
@@ -3,10 +3,10 @@
 
 #![allow(deprecated)]
 
-use crate::error::{ErrorKind, ParseError};
+use crate::error::{ErrMode, ErrorKind, ParseError};
 use crate::input::{AsBytes, Input, ToUsize};
 use crate::lib::std::ops::{AddAssign, Div, Shl, Shr};
-use crate::{ErrMode, IResult};
+use crate::IResult;
 
 /// Generates a parser taking `count` bits
 ///
@@ -30,7 +30,7 @@ use crate::{ErrMode, IResult};
 /// assert_eq!(parser(([0b00010010].as_ref(), 4), 4), Ok((([].as_ref(), 0), 0b00000010)));
 ///
 /// // Tries to consume 12 bits but only 8 are available
-/// assert_eq!(parser(([0b00010010].as_ref(), 0), 12), Err(winnow::ErrMode::Error(Error{input: ([0b00010010].as_ref(), 0), kind: ErrorKind::Eof })));
+/// assert_eq!(parser(([0b00010010].as_ref(), 0), 12), Err(winnow::error::ErrMode::Error(Error{input: ([0b00010010].as_ref(), 0), kind: ErrorKind::Eof })));
 /// ```
 ///
 /// **WARNING:** Deprecated, replaced with [`winnow::bits::take`][crate::bits::take]
@@ -182,7 +182,7 @@ mod test {
 
     assert_eq!(
       result,
-      Err(crate::ErrMode::Error(crate::error::Error {
+      Err(crate::error::ErrMode::Error(crate::error::Error {
         input: (input, 8),
         kind: ErrorKind::Eof
       }))
@@ -218,7 +218,7 @@ mod test {
 
     assert_eq!(
       result,
-      Err(crate::ErrMode::Error(crate::error::Error {
+      Err(crate::error::ErrMode::Error(crate::error::Error {
         input: (input, 8),
         kind: ErrorKind::Eof
       }))
