+++ src/bits/tests.rs
@@ -80,7 +80,7 @@ fn test_take_complete_eof() {
 
   assert_eq!(
     result,
-    Err(crate::ErrMode::Error(crate::error::Error {
+    Err(crate::error::ErrMode::Error(crate::error::Error {
       input: (input, 8),
       kind: ErrorKind::Eof
     }))
@@ -134,7 +134,7 @@ fn test_tag_streaming_err() {
 
   assert_eq!(
     result,
-    Err(crate::ErrMode::Error(crate::error::Error {
+    Err(crate::error::ErrMode::Error(crate::error::Error {
       input: (input, offset),
       kind: ErrorKind::TagBits
     }))
@@ -158,7 +158,7 @@ fn test_bool_eof_complete() {
 
   assert_eq!(
     result,
-    Err(crate::ErrMode::Error(crate::error::Error {
+    Err(crate::error::ErrMode::Error(crate::error::Error {
       input: (input, 8),
       kind: ErrorKind::Eof
     }))
@@ -180,5 +180,8 @@ fn test_bool_eof_streaming() {
 
   let result: crate::IResult<(Streaming<&[u8]>, usize), bool> = bool((input, 8));
 
-  assert_eq!(result, Err(crate::ErrMode::Incomplete(Needed::new(1))));
+  assert_eq!(
+    result,
+    Err(crate::error::ErrMode::Incomplete(Needed::new(1)))
+  );
 }
