+++ src/error.rs
@@ -10,7 +10,7 @@
 //! type:
 //!
 //! ```rust
-//! pub type IResult<I, O, E=winnow::error::Error<I>> = Result<(I, O), winnow::ErrMode<E>>;
+//! pub type IResult<I, O, E=winnow::error::Error<I>> = Result<(I, O), winnow::error::ErrMode<E>>;
 //!
 //! pub enum ErrMode<E> {
 //!     Incomplete(Needed),
@@ -443,7 +443,294 @@
 //! You can go further with the [nom-trace crate](https://github.com/rust-bakery/nom-trace)
 
 use crate::lib::std::fmt;
+use core::num::NonZeroUsize;
+
+use crate::input::Input;
+use crate::input::InputIsStreaming;
 use crate::Parser;
+
+/// Holds the result of parsing functions
+///
+/// It depends on the input type `I`, the output type `O`, and the error type `E`
+/// (by default `(I, winnow::ErrorKind)`)
+///
+/// The `Ok` side is a pair containing the remainder of the input (the part of the data that
+/// was not parsed) and the produced value. The `Err` side contains an instance of `winnow::Err`.
+///
+/// Outside of the parsing code, you can use the [`FinishIResult::finish`] method to convert
+/// it to a more common result type
+pub type IResult<I, O, E = Error<I>> = Result<(I, O), ErrMode<E>>;
+
+/// Extension trait to convert a parser's [`IResult`] to a more manageable type
+pub trait FinishIResult<I, O, E> {
+  /// Converts the parser's [`IResult`] to a type that is more consumable by callers.
+  ///
+  /// Errors if the parser is not at the [end of input][crate::combinator::eof].  See
+  /// [`FinishIResult::finish_err`] if the remaining input is needed.
+  ///
+  /// # Panic
+  ///
+  /// If the result is `Err(ErrMode::Incomplete(_))`, this method will panic.
+  /// - "complete" parsers: It will not be an issue, `Incomplete` is never used
+  /// - "streaming" parsers: `Incomplete` will be returned if there's not enough data
+  /// for the parser to decide, and you should gather more data before parsing again.
+  /// Once the parser returns either `Ok(_)`, `Err(ErrMode::Error(_))` or `Err(ErrMode::Failure(_))`,
+  /// you can get out of the parsing loop and call `finish_err()` on the parser's result
+  fn finish(self) -> Result<O, E>;
+
+  /// Converts the parser's [`IResult`] to a type that is more consumable by errors.
+  ///
+  ///  It keeps the same `Ok` branch, and merges `ErrMode::Error` and `ErrMode::Failure` into the `Err`
+  ///  side.
+  ///
+  /// # Panic
+  ///
+  /// If the result is `Err(ErrMode::Incomplete(_))`, this method will panic as [`ErrMode::Incomplete`]
+  /// should only be set when the input is [`InputIsStreaming<false>`] which this isn't implemented
+  /// for.
+  fn finish_err(self) -> Result<(I, O), E>;
+}
+
+impl<I, O, E> FinishIResult<I, O, E> for IResult<I, O, E>
+where
+  I: Input,
+  // Force users to deal with `Incomplete` when `InputIsStreaming<true>`
+  I: InputIsStreaming<false>,
+  I: Clone,
+  E: ParseError<I>,
+{
+  fn finish(self) -> Result<O, E> {
+    let (i, o) = self.finish_err()?;
+    crate::combinator::eof(i).finish_err()?;
+    Ok(o)
+  }
+
+  fn finish_err(self) -> Result<(I, O), E> {
+    match self {
+      Ok(res) => Ok(res),
+      Err(ErrMode::Error(e)) | Err(ErrMode::Failure(e)) => Err(e),
+      Err(ErrMode::Incomplete(_)) => {
+        panic!("`InputIsStreaming<false>` conflicts with `Err(ErrMode::Incomplete(_))`")
+      }
+    }
+  }
+}
+
+#[doc(hidden)]
+#[deprecated(
+  since = "8.0.0",
+  note = "Replaced with `FinishIResult` which is available via `winnow::prelude`"
+)]
+pub trait Finish<I, O, E> {
+  #[deprecated(
+    since = "8.0.0",
+    note = "Replaced with `FinishIResult::finish_err` which is available via `winnow::prelude`"
+  )]
+  fn finish(self) -> Result<(I, O), E>;
+}
+
+#[allow(deprecated)]
+impl<I, O, E> Finish<I, O, E> for IResult<I, O, E> {
+  fn finish(self) -> Result<(I, O), E> {
+    match self {
+      Ok(res) => Ok(res),
+      Err(ErrMode::Error(e)) | Err(ErrMode::Failure(e)) => Err(e),
+      Err(ErrMode::Incomplete(_)) => {
+        panic!("Cannot call `finish()` on `Err(ErrMode::Incomplete(_))`: this result means that the parser does not have enough data to decide, you should gather more data and try to reapply  the parser instead")
+      }
+    }
+  }
+}
+
+/// Contains information on needed data if a parser returned `Incomplete`
+///
+/// **Note:** This is only possible for `Input` types that implement [`InputIsStreaming<true>`],
+/// like [`Streaming`][crate::input::Streaming].
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+#[cfg_attr(nightly, warn(rustdoc::missing_doc_code_examples))]
+pub enum Needed {
+  /// Needs more data, but we do not know how much
+  Unknown,
+  /// Contains the required data size in bytes
+  Size(NonZeroUsize),
+}
+
+impl Needed {
+  /// Creates `Needed` instance, returns `Needed::Unknown` if the argument is zero
+  pub fn new(s: usize) -> Self {
+    match NonZeroUsize::new(s) {
+      Some(sz) => Needed::Size(sz),
+      None => Needed::Unknown,
+    }
+  }
+
+  /// Indicates if we know how many bytes we need
+  pub fn is_known(&self) -> bool {
+    *self != Needed::Unknown
+  }
+
+  /// Maps a `Needed` to `Needed` by applying a function to a contained `Size` value.
+  #[inline]
+  pub fn map<F: Fn(NonZeroUsize) -> usize>(self, f: F) -> Needed {
+    match self {
+      Needed::Unknown => Needed::Unknown,
+      Needed::Size(n) => Needed::new(f(n)),
+    }
+  }
+}
+
+/// The `Err` enum indicates the parser was not successful
+///
+/// It has three cases:
+///
+/// * `Incomplete` indicates that more data is needed to decide. The [`Needed`] enum
+/// can contain how many additional bytes are necessary. If you are sure your parser
+/// is working on full data, you can wrap your parser with the `complete` combinator
+/// to transform that case in `Error`
+/// * `Error` means some parser did not succeed, but another one might (as an example,
+/// when testing different branches of an `alt` combinator)
+/// * `Failure` indicates an unrecoverable error. As an example, if you recognize a prefix
+/// to decide on the next parser to apply, and that parser fails, you know there's no need
+/// to try other parsers, you were already in the right branch, so the data is invalid
+///
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(nightly, warn(rustdoc::missing_doc_code_examples))]
+pub enum ErrMode<E> {
+  /// There was not enough data
+  ///
+  /// This must only be set when the `Input` is [`InputIsStreaming<true>`], like with
+  /// [`Streaming`][crate::input::Streaming]
+  ///
+  /// Convert this into an `Error` with [`Parser::complete`][Parser::complete]
+  Incomplete(Needed),
+  /// The parser had an error (recoverable)
+  Error(E),
+  /// The parser had an unrecoverable error: we got to the right
+  /// branch and we know other branches won't work, so backtrack
+  /// as fast as possible
+  Failure(E),
+}
+
+impl<E> ErrMode<E> {
+  /// Tests if the result is Incomplete
+  pub fn is_incomplete(&self) -> bool {
+    matches!(self, ErrMode::Incomplete(_))
+  }
+
+  /// Applies the given function to the inner error
+  pub fn map<E2, F>(self, f: F) -> ErrMode<E2>
+  where
+    F: FnOnce(E) -> E2,
+  {
+    match self {
+      ErrMode::Incomplete(n) => ErrMode::Incomplete(n),
+      ErrMode::Failure(t) => ErrMode::Failure(f(t)),
+      ErrMode::Error(t) => ErrMode::Error(f(t)),
+    }
+  }
+
+  /// Automatically converts between errors if the underlying type supports it
+  pub fn convert<F>(e: ErrMode<F>) -> Self
+  where
+    F: Into<E>,
+  {
+    e.map(crate::lib::std::convert::Into::into)
+  }
+}
+
+impl<T> ErrMode<(T, ErrorKind)> {
+  /// Maps `ErrMode<(T, ErrorKind)>` to `ErrMode<(U, ErrorKind)>` with the given `F: T -> U`
+  pub fn map_input<U, F>(self, f: F) -> ErrMode<(U, ErrorKind)>
+  where
+    F: FnOnce(T) -> U,
+  {
+    match self {
+      ErrMode::Incomplete(n) => ErrMode::Incomplete(n),
+      ErrMode::Failure((input, k)) => ErrMode::Failure((f(input), k)),
+      ErrMode::Error((input, k)) => ErrMode::Error((f(input), k)),
+    }
+  }
+}
+
+impl<T> ErrMode<Error<T>> {
+  /// Maps `ErrMode<Error<T>>` to `ErrMode<Error<U>>` with the given `F: T -> U`
+  pub fn map_input<U, F>(self, f: F) -> ErrMode<Error<U>>
+  where
+    F: FnOnce(T) -> U,
+  {
+    match self {
+      ErrMode::Incomplete(n) => ErrMode::Incomplete(n),
+      ErrMode::Failure(Error { input, kind }) => ErrMode::Failure(Error {
+        input: f(input),
+        kind,
+      }),
+      ErrMode::Error(Error { input, kind }) => ErrMode::Error(Error {
+        input: f(input),
+        kind,
+      }),
+    }
+  }
+}
+
+#[cfg(feature = "alloc")]
+use crate::lib::std::{borrow::ToOwned, string::String, vec::Vec};
+impl ErrMode<(&[u8], ErrorKind)> {
+  /// Obtaining ownership
+  #[cfg(feature = "alloc")]
+  pub fn to_owned(self) -> ErrMode<(Vec<u8>, ErrorKind)> {
+    self.map_input(ToOwned::to_owned)
+  }
+}
+
+impl ErrMode<(&str, ErrorKind)> {
+  /// Obtaining ownership
+  #[cfg(feature = "alloc")]
+  pub fn to_owned(self) -> ErrMode<(String, ErrorKind)> {
+    self.map_input(ToOwned::to_owned)
+  }
+}
+
+impl ErrMode<Error<&[u8]>> {
+  /// Obtaining ownership
+  #[cfg(feature = "alloc")]
+  pub fn to_owned(self) -> ErrMode<Error<Vec<u8>>> {
+    self.map_input(ToOwned::to_owned)
+  }
+}
+
+impl ErrMode<Error<&str>> {
+  /// Obtaining ownership
+  #[cfg(feature = "alloc")]
+  pub fn to_owned(self) -> ErrMode<Error<String>> {
+    self.map_input(ToOwned::to_owned)
+  }
+}
+
+impl<E: Eq> Eq for ErrMode<E> {}
+
+impl<E> fmt::Display for ErrMode<E>
+where
+  E: fmt::Debug,
+{
+  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+    match self {
+      ErrMode::Incomplete(Needed::Size(u)) => write!(f, "Parsing requires {} bytes/chars", u),
+      ErrMode::Incomplete(Needed::Unknown) => write!(f, "Parsing requires more data"),
+      ErrMode::Failure(c) => write!(f, "Parsing Failure: {:?}", c),
+      ErrMode::Error(c) => write!(f, "Parsing Error: {:?}", c),
+    }
+  }
+}
+
+#[cfg(feature = "std")]
+impl<E> std::error::Error for ErrMode<E>
+where
+  E: fmt::Debug,
+{
+  fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+    None // no underlying error
+  }
+}
 
 /// This trait must be implemented by the error type of a nom parser.
 ///
@@ -680,8 +967,6 @@ impl<I: fmt::Display> fmt::Display for VerboseError<I> {
 
 #[cfg(feature = "std")]
 impl<I: fmt::Debug + fmt::Display> std::error::Error for VerboseError<I> {}
-
-use crate::{ErrMode, IResult};
 
 /// Create a new error from an input position, a static string and an existing error.
 /// This is used mainly in the [context] combinator, to add user friendly information
